refactor: separate drift package responsibilities
